backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.25
alembic>=1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
twilio==8.10.0
stripe==7.7.0
crewai
openai
apscheduler==3.10.4
python-dotenv==1.0.0
httpx==0.25.2